requirements.txt
fastapi
uvicorn
transformers
faiss-cpu
psycopg2-binary
torch
pandas
numpy
